+++ src/libtropic_l3.c
@@ -16,6 +16,7 @@
 #include "libtropic.h"
 #include "libtropic_common.h"
 #include "libtropic_l2.h"
+#include "libtropic_logging.h"
 #include "libtropic_port.h"
 #include "libtropic_secure_memzero.h"
 #include "lt_aesgcm.h"
@@ -1148,6 +1149,7 @@ lt_ret_t lt_in__ecc_key_read(lt_handle_t *h, uint8_t *key, const uint8_t key_max
     }
     else {
         // Unknown curve type.
+        LT_LOG_ERROR("Unknown curve type in ECC_Key_Read Result!");
         return LT_FAIL;
     }